moved the cohere file to proper folder and deleted old file
